refactor: fail to components editor
